+++ src/digsim/app/gui_objects/_yosys_object.py
@@ -5,6 +5,8 @@
 
 from PySide6.QtGui import QAction
 
+from digsim.circuit.components.atoms import DigsimException
+
 from ._image_objects import ImageObject
 
 
@@ -23,5 +25,8 @@ class YosysObject(ImageObject):
         reloadAction.triggered.connect(self._reload)
 
     def _reload(self):
-        self.component.reload_file()
+        try:
+            self.component.reload_file()
+        except DigsimException as exc:
+            self._app_model.sig_warning_log.emit("Reload Yosys Warning", str(exc))
         self._app_model.model_reset()