+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## vx.x.x
  * Fix bug when detecting number of modules with yosys 0.9
+ * Improve error handling for yosys component
 
 ## v0.0.1 - First Release
  * Simulation of a digital circuit using a python